\name{accumulator-package}
\alias{accumulator-package}
\alias{accumulator}
\docType{package}
\title{
\packageTitle{accumulator}
}
\description{
\packageDescription{accumulator}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{accumulator}

Maintainer: \packageMaintainer{accumulator}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = accumulator)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{accumulator}}
%\section{Documentation Index}{\packageIndices{accumulator}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
